pyproject.toml
[project]
name = "small_size_league_promoter"
version = "0.1.0"
description = "RoboCup SSL article generation system using CrewAI"
authors = [{ name = "Bruno Castro", email = "bruno.c0310@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]==0.121.0",
    "langchain-groq>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "fastapi[standard]>=0.109.0",
    "uvicorn>=0.27.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "langchain-openai>=0.2.14",
    "crewai-tools[mcp]>=0.40.1",
    "mcp>=1.6.0",
    "ruff>=0.11.6",
    "discord.py>=2.0.0",
    "fastmcp>=2.3",
]

[project.scripts]
run_crew = "small_size_league_promoter.main:run"
run_api = "small_size_league_promoter.main:run_api"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
